+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/FormatExtension.java
@@ -35,6 +35,8 @@ import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.LazyForwardingEquality;
 import com.diffplug.spotless.LineEnding;
 import com.diffplug.spotless.ThrowingEx;
+import com.diffplug.spotless.extra.EclipseBasedStepBuilder;
+import com.diffplug.spotless.extra.wtp.EclipseWtpFormatterStep;
 import com.diffplug.spotless.generic.EndWithNewlineStep;
 import com.diffplug.spotless.generic.IndentStep;
 import com.diffplug.spotless.generic.LicenseHeaderStep;
@@ -481,6 +483,31 @@ public class FormatExtension {
 		final PrettierConfig prettierConfig = new PrettierConfig();
 		addStep(prettierConfig.createStep());
 		return prettierConfig;
+	}
+
+	public class EclipseWtpConfig {
+		private final EclipseBasedStepBuilder builder;
+
+		EclipseWtpConfig(EclipseWtpFormatterStep type, String version) {
+			builder = type.createBuilder(GradleProvisioner.fromProject(getProject()));
+			builder.setVersion(version);
+			addStep(builder.build());
+		}
+
+		public void configFile(Object... configFiles) {
+			requireElementsNonNull(configFiles);
+			Project project = getProject();
+			builder.setPreferences(project.files(configFiles).getFiles());
+			replaceStep(builder.build());
+		}
+	}
+
+	public EclipseWtpConfig eclipseWtp(EclipseWtpFormatterStep type) {
+		return eclipseWtp(type, EclipseWtpFormatterStep.defaultVersion());
+	}
+
+	public EclipseWtpConfig eclipseWtp(EclipseWtpFormatterStep type, String version) {
+		return new EclipseWtpConfig(type, version);
 	}
 
 	/** Sets up a format task according to the values in this extension. */